requirements.txt
discord.py
selenium
beautifulsoup4
webdriver-manager
python-dotenv
gspread
google-auth-oauthlib
pandas